requirements.txt
requests
pandas
numpy
yfinance
pandas_ta
apscheduler
python-dotenv
pytz
